+++ jwt-servlet-auth-core/src/test/java/io/telicent/servlet/auth/jwt/fake/TestFakeConfigurableFilter.java
@@ -95,7 +95,7 @@ public class TestFakeConfigurableFilter
     protected JwtAuthenticationEngine<FakeRequest, FakeResponse> createEngine(List<HeaderSource> authHeaders,
                                                                               String realm,
                                                                               List<String> usernameClaims) {
-        return new FakeEngine(authHeaders, realm, usernameClaims);
+        return new FakeEngine(authHeaders, realm, usernameClaims, null);
     }
 
     @Override