+++ jwt-servlet-auth-jaxrs3/src/main/java/io/telicent/servlet/auth/jwt/jaxrs3/JaxRs3EngineProvider.java
@@ -28,9 +28,10 @@ public class JaxRs3EngineProvider extends AbstractHeaderBasedEngineProvider {
     @Override
     @SuppressWarnings("unchecked")
     protected <TRequest, TResponse> JwtAuthenticationEngine<TRequest, TResponse> createEngine(
-            List<HeaderSource> headerSources, String realm, List<String> usernameClaims) {
+            List<HeaderSource> headerSources, String realm, List<String> usernameClaims, String[] rolesClaim) {
         return (JwtAuthenticationEngine<TRequest, TResponse>) new JaxRs3JwtAuthenticationEngine(headerSources, realm,
-                                                                                                usernameClaims);
+                                                                                                usernameClaims,
+                                                                                                rolesClaim);
     }
 
     @Override