+++ jwt-servlet-auth-jaxrs3/src/test/java/io/telicent/servlet/auth/jwt/jaxrs3/TestJaxRs3Engine.java
@@ -53,11 +53,11 @@ public class TestJaxRs3Engine
     }
 
     @Override
-    @SuppressWarnings("unchecked")
     protected ContainerRequestContext createMockRequest(Map<String, String> headers) {
         return mockRequest(TEST_REQUEST_URI, headers);
     }
 
+    @SuppressWarnings("unchecked")
     public static ContainerRequestContext mockRequest(URI requestUri, Map<String, String> headers) {
         ContainerRequestContext request = mock(ContainerRequestContext.class);
         MultivaluedMap<String, String> mockHeaders = mock(MultivaluedMap.class);
@@ -130,13 +130,13 @@ public class TestJaxRs3Engine
                                                                                                       String realm,
                                                                                                       String usernameClaim) {
         return new JaxRs3JwtAuthenticationEngine(List.of(new HeaderSource(authHeader, authHeaderPrefix)), realm,
-                                                 usernameClaim != null ? List.of(usernameClaim) : null);
+                                                 usernameClaim != null ? List.of(usernameClaim) : null, null);
     }
 
     @Override
     protected JwtAuthenticationEngine<ContainerRequestContext, ContainerResponseContext> createEngine(
             List<HeaderSource> authHeaders, String realm, List<String> usernameClaims) {
-        return new JaxRs3JwtAuthenticationEngine(authHeaders, realm, usernameClaims);
+        return new JaxRs3JwtAuthenticationEngine(authHeaders, realm, usernameClaims, null);
     }
 
     @Override
@@ -150,7 +150,7 @@ public class TestJaxRs3Engine
         verifyStatusCode(request, expectedStatus);
     }
 
-    public final static void verifyStatusCode(ContainerRequestContext request, int expectedStatus) {
+    public static void verifyStatusCode(ContainerRequestContext request, int expectedStatus) {
         ArgumentCaptor<Response> captor = ArgumentCaptor.forClass(Response.class);
         verify(request).abortWith(captor.capture());
         Response actualResponse = captor.getValue();
@@ -163,7 +163,7 @@ public class TestJaxRs3Engine
         return verifyHeaderPresent(request, expectedHeader);
     }
 
-    public final static String verifyHeaderPresent(ContainerRequestContext request, String expectedHeader) {
+    public static String verifyHeaderPresent(ContainerRequestContext request, String expectedHeader) {
         ArgumentCaptor<Response> captor = ArgumentCaptor.forClass(Response.class);
         verify(request).abortWith(captor.capture());
         Response actualResponse = captor.getValue();
@@ -178,7 +178,7 @@ public class TestJaxRs3Engine
         return verifyAuthenticatedUser(authenticatedRequest);
     }
 
-    public final static String verifyAuthenticatedUser(ContainerRequestContext authenticatedRequest) {
+    public static String verifyAuthenticatedUser(ContainerRequestContext authenticatedRequest) {
         ArgumentCaptor<SecurityContext> captor = ArgumentCaptor.forClass(SecurityContext.class);
         verify(authenticatedRequest).setSecurityContext(captor.capture());
         SecurityContext context = captor.getValue();