+++ jwt-servlet-auth-servlet3/src/test/java/io/telicent/servlet/auth/jwt/servlet3/TestServlet3Filter.java
@@ -112,13 +112,13 @@ public class TestServlet3Filter extends
                                                                                             String realm,
                                                                                             String usernameClaim) {
         return new Servlet3JwtAuthenticationEngine(List.of(new HeaderSource(authHeader, authHeaderPrefix)), realm,
-                                                   usernameClaim != null ? List.of(usernameClaim) : null);
+                                                   usernameClaim != null ? List.of(usernameClaim) : null, null);
     }
 
     @Override
     protected JwtAuthenticationEngine<HttpServletRequest, HttpServletResponse> createEngine(
             List<HeaderSource> authHeaders, String realm, List<String> usernameClaims) {
-        return new Servlet3JwtAuthenticationEngine(authHeaders, realm, usernameClaims);
+        return new Servlet3JwtAuthenticationEngine(authHeaders, realm, usernameClaims, null);
     }
 
     @Override